Refactoring Firebase code and small cleanup.
